v0.8.0 Improvement (#345) - removed startingchallenge button
